requirements.txt
PyYAML
pydantic>=2.0,<3.0
# dev
pytest
pytest-mock
coverage
# temporary
fastapi
uvicorn
pyqt
